fix: remove some error codes and unused imports
